add movie GET route - fetch movies with pagination from mongodb - refactor home page - add favicon
